# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Rawaa API",
    "version": "1.0.0",
    "description": "API documentation for Rawaa project"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local server"
    }
  ],
  "paths": {
    "/rawa/ads": {
      "post": {
        "summary": "Create a new advertisement",
        "tags": [
          "Ads"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "Only logged-in users with roles SUPER_ADMIN, ADMIN, or DATA_ENTRY can create a new ad. Upload one image (required). The expiration date is optional; defaults to 7 days if not provided.\n",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "photo": {
                    "type": "string",
                    "format": "binary"
                  },
                  "expiresAt": {
                    "type": "string",
                    "format": "date-time"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Advertisement created successfully"
          },
          "400": {
            "description": "Missing photo or invalid request"
          },
          "401": {
            "description": "Unauthorized – user not logged in"
          },
          "403": {
            "description": "Forbidden – user does not have required role"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "get": {
        "summary": "Get all ads",
        "tags": [
          "Ads"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "Retrieve all ads. Only authorized 'SUPER_ADMIN', 'ADMIN', 'DATA_ENTRY' can access.",
        "responses": {
          "200": {
            "description": "List of ads"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/rawa/ads/{id}/show": {
      "patch": {
        "summary": "Show an ad",
        "tags": [
          "Ads"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "Only SUPER_ADMIN, ADMIN, DATA_ENTRY can make an ad visible.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "description": "Advertisement ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Ad is now visible"
          },
          "404": {
            "description": "Ad not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/rawa/ads/{id}/hide": {
      "patch": {
        "summary": "Hide ad manually",
        "tags": [
          "Ads"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "Only SUPER_ADMIN, ADMIN, DATA_ENTRY can hide an ad.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Ad ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Ad is now hidden"
          },
          "404": {
            "description": "Ad not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/rawa/ads/active/showing": {
      "get": {
        "summary": "Get all currently active & visible ads",
        "tags": [
          "Ads"
        ],
        "description": "Returns ads with inShow = true and not expired. Expired ads will be hidden automatically.",
        "responses": {
          "200": {
            "description": "List of visible ads"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/rawa/ads/{id}": {
      "get": {
        "summary": "Get ad by ID",
        "tags": [
          "Ads"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "Retrieve all ads. Only authorized 'SUPER_ADMIN', 'ADMIN', 'DATA_ENTRY' can access.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "description": "Ad ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Ad found"
          },
          "404": {
            "description": "Ad not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "put": {
        "summary": "Update an existing ad",
        "tags": [
          "Ads"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "Retrieve all ads. Only authorized 'SUPER_ADMIN', 'ADMIN', 'DATA_ENTRY' can access.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "description": "Ad ID"
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "photo": {
                    "type": "string",
                    "format": "binary"
                  },
                  "expiresAt": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "inShow": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Ad updated successfully"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Ad not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "summary": "Delete an ad",
        "tags": [
          "Ads"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "Retrieve all ads. Only authorized 'SUPER_ADMIN', 'ADMIN', 'DATA_ENTRY' can access.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "description": "Ad ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Ad deleted successfully"
          },
          "404": {
            "description": "Ad not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/rawa/signup": {
      "post": {
        "summary": "Register a new user",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string",
                    "example": "johndoe"
                  },
                  "password": {
                    "type": "string",
                    "example": "mySecret123"
                  },
                  "email": {
                    "type": "string",
                    "example": "john@example.com"
                  },
                  "phone": {
                    "type": "string",
                    "example": "+1234567890"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully (may need email verification)"
          },
          "400": {
            "description": "Validation error"
          },
          "409": {
            "description": "Username or email already exists"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/rawa/login": {
      "post": {
        "summary": "Login with username or email",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "identifier": {
                    "type": "string",
                    "description": "username or email",
                    "example": "johndoe"
                  },
                  "password": {
                    "type": "string",
                    "example": "mySecret123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully logged in"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Invalid credentials"
          },
          "403": {
            "description": "Email not verified"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/rawa/verify-email": {
      "post": {
        "summary": "Verify user's email using code",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "john@example.com"
                  },
                  "code": {
                    "type": "string",
                    "example": "123456"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Email verified successfully, JWT token returned"
          },
          "400": {
            "description": "Invalid code or missing parameters"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/rawa/resend-code": {
      "post": {
        "summary": "Resend verification code to user's email",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "john@example.com"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Verification code sent"
          },
          "400": {
            "description": "No email to verify"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/rawa/user/forgot-password": {
      "post": {
        "summary": "Send a password reset code to email",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "john@example.com"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Verification code sent"
          },
          "400": {
            "description": "Validation error"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/rawa/user/reset-password": {
      "post": {
        "summary": "Reset password using code",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "john@example.com"
                  },
                  "code": {
                    "type": "string",
                    "example": "123456"
                  },
                  "newPassword": {
                    "type": "string",
                    "example": "myNewPass123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password reset successfully"
          },
          "400": {
            "description": "Invalid code or validation error"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/rawa/user/change-password/{id}": {
      "post": {
        "summary": "Change password (authenticated)",
        "tags": [
          "Auth"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "User ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "oldPassword": {
                    "type": "string",
                    "example": "oldPass123"
                  },
                  "newPassword": {
                    "type": "string",
                    "example": "newPass123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password changed successfully"
          },
          "400": {
            "description": "Validation error or old password incorrect"
          },
          "403": {
            "description": "Forbidden – not allowed to change this user's password"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    }
  },
  "components": {},
  "tags": []
}